+++ exonum-java-binding-core/rust/integration_tests/tests/node_context.rs
@@ -3,26 +3,25 @@ extern crate integration_tests;
 extern crate java_bindings;
 #[macro_use]
 extern crate lazy_static;
+extern crate failure;
 
 use std::sync::Arc;
 
-use futures::Stream;
-use futures::sync::mpsc::{self, Receiver};
-
+use futures::{
+    sync::mpsc::{self, Receiver},
+    Stream,
+};
 use integration_tests::vm::create_vm_for_tests_with_fake_classes;
 use java_bindings::{
-    Java_com_exonum_binding_service_NodeProxy_nativeSubmit, JniExecutor, JniResult, MainExecutor,
-    NodeContext,
-};
-use java_bindings::exonum::blockchain::Blockchain;
-use java_bindings::exonum::crypto::gen_keypair;
-use java_bindings::exonum::messages::{BinaryForm, RawTransaction, ServiceTransaction};
-use java_bindings::exonum::node::{ApiSender, ExternalMessage};
-use java_bindings::exonum::storage::MemoryDB;
-use java_bindings::jni::{JavaVM, JNIEnv};
-use java_bindings::jni::objects::JObject;
-use java_bindings::utils::{
-    as_handle, get_and_clear_java_exception, get_class_name, unwrap_jni, unwrap_jni_verbose,
+    exonum::{
+        blockchain::{Blockchain, Service, Transaction},
+        crypto::{gen_keypair, Hash, PublicKey, SecretKey},
+        messages::{RawTransaction, ServiceTransaction},
+        node::{ApiSender, ExternalMessage},
+        storage::{MemoryDB, Snapshot},
+    },
+    jni::JavaVM,
+    MainExecutor, NodeContext,
 };
 
 lazy_static! {
@@ -33,12 +32,14 @@ lazy_static! {
 (extracts some variables like service_id, signs the transaction) can we test that as well? */
 
 #[test]
-//TODO
-#[ignore]
 fn submit_transaction() {
-    let (mut node, app_rx) = create_node();
-    let raw_transaction =
-        RawTransaction::new(0, ServiceTransaction::from_raw_unchecked(0, vec![1, 2, 3]));
+    let keypair = gen_keypair();
+    let (node, app_rx) = create_node(keypair.0, keypair.1);
+    let service_id = 0;
+    let transaction_id = 0;
+    let tx_payload = vec![1, 2, 3];
+    let service_transaction = ServiceTransaction::from_raw_unchecked(transaction_id, tx_payload);
+    let raw_transaction = RawTransaction::new(service_id, service_transaction);
     node.submit(raw_transaction.clone()).unwrap();
     let sent_message = app_rx.wait().next().unwrap().unwrap();
     /*
@@ -47,28 +48,65 @@ fn submit_transaction() {
     setn.payload in a variable.
     */
     match sent_message {
-        ExternalMessage::Transaction(sent) => assert_eq!(&raw_transaction, sent.payload()),
+        ExternalMessage::Transaction(sent) => {
+            let tx_payload = sent.payload();
+            let tx_author = sent.author();
+            assert_eq!(&raw_transaction, tx_payload);
+            assert_eq!(tx_author, keypair.0);
+        }
         _ => panic!("Message is not Transaction"),
     }
 }
 
-fn create_node() -> (NodeContext, Receiver<ExternalMessage>) {
-    let service_keypair = gen_keypair();
+#[test]
+fn submit_transaction_to_missing_service() {
+    let keypair = gen_keypair();
+    let (node, _) = create_node(keypair.0, keypair.1);
+    // invalid service_id
+    let service_id = 1;
+    let transaction_id = 0;
+    let tx_payload = vec![1, 2, 3];
+    let service_transaction = ServiceTransaction::from_raw_unchecked(transaction_id, tx_payload);
+    let raw_transaction = RawTransaction::new(service_id, service_transaction);
+    let res = node.submit(raw_transaction.clone());
+    assert!(res.is_err());
+}
+
+fn create_node(
+    public_key: PublicKey,
+    secret_key: SecretKey,
+) -> (NodeContext, Receiver<ExternalMessage>) {
     let api_channel = mpsc::channel(128);
     let (app_tx, app_rx) = (ApiSender::new(api_channel.0), api_channel.1);
+
+    struct EmptyService;
+
+    impl Service for EmptyService {
+        fn service_id(&self) -> u16 {
+            0
+        }
+
+        fn service_name(&self) -> &str {
+            "empty_service"
+        }
+
+        fn state_hash(&self, _: &Snapshot) -> Vec<Hash> {
+            vec![]
+        }
+
+        fn tx_from_raw(&self, _: RawTransaction) -> Result<Box<dyn Transaction>, failure::Error> {
+            unimplemented!()
+        }
+    }
 
     let storage = MemoryDB::new();
     let blockchain = Blockchain::new(
         storage,
-        vec![],
-        service_keypair.0,
-        service_keypair.1,
+        vec![Box::new(EmptyService)],
+        public_key,
+        secret_key,
         app_tx.clone(),
     );
-    let node = NodeContext::new(EXECUTOR.clone(), blockchain, service_keypair.0, app_tx);
+    let node = NodeContext::new(EXECUTOR.clone(), blockchain, public_key, app_tx);
     (node, app_rx)
-}
-
-fn message_from_raw<'e>(env: &'e JNIEnv<'e>, buffer: &[u8]) -> JniResult<JObject<'e>> {
-    env.byte_array_from_slice(buffer).map(JObject::from)
 }