+++ exonum-java-binding-core/rust/integration_tests/tests/transaction_proxy.rs
@@ -3,27 +3,27 @@ extern crate java_bindings;
 #[macro_use]
 extern crate lazy_static;
 
-use std::sync::Arc;
-
 use integration_tests::{
     mock::transaction::{
-        AUTHOR_PK_ENTRY_NAME, create_empty_raw_transaction,
-        create_mock_transaction_proxy,
-        create_throwing_exec_exception_mock_transaction_proxy, create_throwing_mock_transaction_proxy, ENTRY_VALUE, INFO_VALUE,
-        TEST_ENTRY_NAME, TX_HASH_ENTRY_NAME,
+        create_mock_transaction_proxy, create_throwing_exec_exception_mock_transaction_proxy,
+        create_throwing_mock_transaction_proxy, AUTHOR_PK_ENTRY_NAME, ENTRY_VALUE, TEST_ENTRY_NAME,
+        TX_HASH_ENTRY_NAME,
     },
     vm::create_vm_for_tests_with_fake_classes,
 };
+
 use java_bindings::{
     exonum::{
         blockchain::{Transaction, TransactionContext, TransactionError, TransactionErrorType},
         crypto::{Hash, PublicKey},
-        messages::{Message, RawTransaction, ServiceTransaction},
+        messages::RawTransaction,
         storage::{Database, Entry, Fork, MemoryDB, Snapshot},
     },
     jni::JavaVM,
     MainExecutor,
 };
+
+use std::sync::Arc;
 
 const ARITHMETIC_EXCEPTION_CLASS: &str = "java/lang/ArithmeticException";
 const OOM_ERROR_CLASS: &str = "java/lang/OutOfMemoryError";
@@ -43,7 +43,7 @@ fn execute_valid_transaction() {
     }
     {
         let mut fork = db.fork();
-        let (valid_tx, raw) = create_mock_transaction_proxy(EXECUTOR.clone(), true);
+        let (valid_tx, raw) = create_mock_transaction_proxy(EXECUTOR.clone());
         {
             let tx_context = create_transaction_context(&mut fork, raw);
             valid_tx
@@ -177,11 +177,15 @@ fn passing_transaction_context() {
     let db = MemoryDB::new();
     let (tx_hash, author_pk) = {
         let mut fork = db.fork();
-        let (valid_tx, raw) = create_mock_transaction_proxy(EXECUTOR.clone(), true);
+        let (valid_tx, raw) = create_mock_transaction_proxy(EXECUTOR.clone());
+        // get transaction hash and author public key from mock transaction
         let (tx_hash, author_pk) = {
             let context = create_transaction_context(&mut fork, raw);
             let (tx_hash, author_pk) = (context.tx_hash(), context.author());
+
+            // execute transaction
             valid_tx.execute(context).unwrap();
+
             (tx_hash, author_pk)
         };
         db.merge(fork.into_patch()).unwrap();