+++ exonum-java-binding-core/src/main/java/com/exonum/binding/transaction/Transaction.java
@@ -44,21 +44,13 @@ public interface Transaction {
    */
   void execute(TransactionContext context) throws TransactionExecutionException;
 
-  /**
-   * Returns a hash of this transaction — a SHA-256 hash of the transaction message.
-   *
-   * @implSpec Default implementation returns {@code getRawTransaction().hash()}.
-   */
-  default HashCode hash() {
-    return getRawTransaction().hash();
-  }
-
-  /**
-   * Returns this transaction as a binary Exonum message.
-   */
   /*
 Review: Is this method really required in each implementation
 Shan't we separate messages from transactions?
   */
-  RawTransaction getRawTransaction();
+  /**
+   * Returns a hash of this transaction — a SHA-256 hash of the transaction message.
+   */
+  HashCode hash();
+
 }